requirements.txt
torch==1.6.0
torchvision==0.7.0
numpy==1.19.5
pillow
tqdm==4.60.0
pandas==1.0.5
